backend/requirements.txt
# Core LLM + Agent
# Unpinned to allow pip to resolve compatible versions automatically.
# Tested with langchain 0.3.x, langgraph 0.2.x, langchain-ollama 0.2.x
langchain>=0.3.7
langchain-ollama>=0.2.1
langchain-community>=0.3.7
langgraph>=0.2.45

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0

# Kubernetes
kubernetes>=29.0.0

# Database + Vector Store
# SQLAlchemy left unpinned — langchain-community has a strict upper bound
# that conflicts with ==2.0.36. Let pip pick a compatible version.
psycopg2-binary>=2.9.0
pgvector>=0.3.0
sqlalchemy>=1.4,<2.0.36

# Document parsing (Phase 2) — Python 3.12 compatible
# unstructured replaced with lighter libraries that work cleanly on 3.12.
# pypdf handles PDF text extraction. markdown-it-py handles markdown parsing.
pypdf>=4.0.0
markdown-it-py>=3.0.0

# Embeddings (via Ollama — no extra pip needed)
# Pull before air-gapping: ollama pull nomic-embed-text

# Utilities
httpx>=0.27.0
tenacity>=9.0.0
rich>=13.0.0
psutil>=5.9
